fix disappearing fees from Cycle overwrite
fix a bug where if a Cycle is created by a withdraw transaction and then a contribution is made in the same cycle, its fees are now vulnerable to be overwritten by a contribution in a subsequent cycle, because it will treat the cycle as having no contributions. If this happens, the overwritten fees won't be processed and distributed to existing participants when they withdraw.